[api] Update iw command, fix 2.4ghz
